Refactor node abstracts
